+++ R/legend.R
@@ -164,9 +164,16 @@ addLegendImage <- function(map,
 #'
 makeSymbol <- function(shape, width, height, color, fillColor = color,
                        opacity = 1, fillOpacity = opacity, ...) {
+  strokewidth <- 0
+  if ( 'stroke-width' %in% names(list(...)) ) {
+    strokewidth <- list(...)[['stroke-width']]
+  }
   shapeTag <- switch(
     shape,
     'rect' = htmltools::tags$rect(
+      id = 'rect',
+      x = strokewidth,
+      y = strokewidth,
       height = height,
       width = width,
       stroke = color,
@@ -176,8 +183,9 @@ makeSymbol <- function(shape, width, height, color, fillColor = color,
       ...
     ),
     'circle' = htmltools::tags$circle(
-      cx = height / 2,
-      cy = height / 2,
+      id = 'circle',
+      cx = height / 2 + strokewidth,
+      cy = height / 2 + strokewidth,
       r = height / 2,
       stroke = color,
       fill = fillColor,
@@ -186,7 +194,52 @@ makeSymbol <- function(shape, width, height, color, fillColor = color,
       ...
     ),
     'triangle' = htmltools::tags$polygon(
-      points = sprintf('0,%d %d,%d %d,0', height, width, height, width / 2),
+      id = 'triangle',
+      points = sprintf('%s,%s %s,%s %s,%s',
+                       strokewidth,
+                       height + strokewidth,
+                       width + strokewidth,
+                       height + strokewidth,
+                       width / 2  + strokewidth,
+                       strokewidth),
+      stroke = color,
+      fill = fillColor,
+      opacity = opacity,
+      'fill-opacity' = fillOpacity,
+      ...
+    ),
+    'plus' = htmltools::tags$polygon(
+      id = 'plus',
+      points = draw_plus(width = width, height = height, offset = strokewidth),
+      stroke = color,
+      fill = fillColor,
+      opacity = opacity,
+      'fill-opacity' = fillOpacity,
+      ...
+    ),
+    'cross' = htmltools::tags$polygon(
+      id = 'cross',
+      points = draw_cross(width = width, height = height, offset = strokewidth),
+      stroke = color,
+      fill = fillColor,
+      opacity = opacity,
+      'fill-opacity' = fillOpacity,
+      ...
+    ),
+    'diamond' = htmltools::tags$polygon(
+      id = 'diamond',
+      points = draw_diamond(width = width, height = height, offset = strokewidth),
+      stroke = color,
+      fill = fillColor,
+      opacity = opacity,
+      'fill-opacity' = fillOpacity,
+      ...
+    ),
+    'star' = htmltools::tags$path(
+      id = 'star',
+      points = sprintf('M%s%z M%sz',
+                      draw_plus(width = width, height = height, offset = strokewidth),
+                      draw_cross(width = width, height = height, offset = strokewidth)),
       stroke = color,
       fill = fillColor,
       opacity = opacity,
@@ -194,6 +247,9 @@ makeSymbol <- function(shape, width, height, color, fillColor = color,
       ...
     ),
     'stadium' = htmltools::tags$rect(
+      id = 'stadium',
+      x = strokewidth,
+      y = strokewidth,
       height = height,
       width = width,
       rx = "25%",
@@ -211,8 +267,8 @@ makeSymbol <- function(shape, width, height, color, fillColor = color,
                              htmltools::tags$svg(
                                xmlns = 'http://www.w3.org/2000/svg',
                                version = '1.1',
-                               width = width,
-                               height = height,
+                               width = width + strokewidth * 2,
+                               height = height + strokewidth * 2,
                                shapeTag
                              )
                            )))
@@ -224,6 +280,25 @@ makeLegendSymbol <- function(label, labelStyle, ...) {
     htmltools::tags$img(src = shapeTag, style = "vertical-align: middle; padding: 1px;"),
     htmltools::tags$span(label, style = sprintf("vertical-align: middle; padding: 1px; %s", labelStyle))
   )
+}
+
+draw_plus <- function(width, height, offset = 0) {
+  x <- width * c(rep(c(.4, 0, .4, .6, 1, .6), each = 2), .4) + offset
+  y <- height * c(0, rep(c(.4, .6, 1, .6, .4, 0), each = 2)) + offset
+  paste(x, y, sep = ',', collapse = ' ')
+}
+
+draw_diamond <- function(width, height, offset = 0) {
+  x <- width * c( .5, 0, .5, 1, .5) + offset
+  y <- height * c(0, .5, 1, .5, 0) + offset
+  paste(x, y, sep = ',', collapse = ' ')
+}
+
+draw_cross <- function(width, height, offset = 0) {
+  a <- sqrt(2) / 10
+  x <- width *  c(a, 0, .5 - a, 0, a, .5, 1 - a, 1, .5 + a, 1, 1 - a, .5, a) + offset
+  y <- height * c(0, a, .5, 1 - a, 1, .5 + a, 1, 1 - a, .5, a, 0, .5 - a, 0) + offset
+  paste(x, y, sep = ',', collapse = ' ')
 }
 
 #' Add Customizable Color Legends to a 'leaflet' map widget
@@ -289,6 +364,10 @@ makeLegendSymbol <- function(label, labelStyle, ...) {
 #' @param decreasing
 #'
 #' order of numbers in the legend
+#'
+#' @param opacity
+#'
+#' opacity of the legend items
 #'
 #' @param ...
 #'
@@ -434,6 +513,7 @@ addLegendNumeric <- function(map,
                              tickLength = 4,
                              tickWidth = 1,
                              decreasing = FALSE,
+                             opacity = 1,
                              ...) {
   stopifnot( attr(pal, 'colorType') == 'numeric' )
   rng <- range(values, na.rm = TRUE)
@@ -512,6 +592,7 @@ addLegendNumeric <- function(map,
                            htmltools::tags$def(
                              htmltools::tags$linearGradient(
                                id = id,
+                               opacity = opacity,
                                x1 = x1, y1 = y1, x2 = x2, y2 = y2,
                                htmltools::tagList(Map(htmltools::tags$stop,
                                                       offset = offsets,
@@ -561,11 +642,12 @@ addLegendQuantile <- function(map,
                               values,
                               title = NULL,
                               labelStyle = '',
-                              shape = c('rect', 'circle', 'triangle', 'stadium'),
+                              shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium'),
                               orientation = c('vertical', 'horizontal'),
                               width = 24,
                               height = 24,
                               numberFormat = function(x) {prettyNum(x, big.mark = ',', scientific = FALSE, digits = 1)},
+                              opacity = 1,
                               ...) {
   stopifnot( attr(pal, 'colorType') == 'quantile' )
   shape <- match.arg(shape)
@@ -592,7 +674,8 @@ addLegendQuantile <- function(map,
     color = colors,
     labelStyle = labelStyle,
     height = height,
-    width = width
+    width = width,
+    opacity = opacity
   )
   orientation <- match.arg(orientation)
   if ( orientation == 'vertical' ) {
@@ -614,10 +697,11 @@ addLegendBin <- function(map,
                          values,
                          title = NULL,
                          labelStyle = '',
-                         shape = c('rect', 'circle', 'triangle', 'stadium'),
+                         shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium'),
                          orientation = c('vertical', 'horizontal'),
                          width = 24,
                          height = 24,
+                         opacity = 1,
                          ...) {
   stopifnot( attr(pal, 'colorType') == 'bin' )
   shape <- match.arg(shape)
@@ -630,7 +714,8 @@ addLegendBin <- function(map,
       color = colors,
       labelStyle = labelStyle,
       height = height,
-      width = width)
+      width = width,
+      opacity = opacity)
   orientation <- match.arg(orientation)
   if ( orientation == 'vertical' ) {
     htmlElements <- lapply(htmlElements, htmltools::tagList, htmltools::tags$br())
@@ -651,10 +736,11 @@ addLegendFactor <- function(map,
                             values,
                             title = NULL,
                             labelStyle = '',
-                            shape = c('rect', 'circle', 'triangle', 'stadium'),
+                            shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium'),
                             orientation = c('vertical', 'horizontal'),
                             width = 24,
                             height = 24,
+                            opacity = 1,
                             ...) {
   stopifnot( attr(pal, 'colorType') == 'factor' )
   shape <- match.arg(shape)
@@ -666,7 +752,8 @@ addLegendFactor <- function(map,
       color = colors,
       labelStyle = labelStyle,
       height = height,
-      width = width)
+      width = width,
+      opacity = opacity)
   orientation <- match.arg(orientation)
   if ( orientation == 'vertical' ) {
     htmlElements <- lapply(htmlElements, htmltools::tagList, htmltools::tags$br())